+++ ui/src/components/github-link-target.ts
@@ -13,7 +13,7 @@ export type GitHubLinkTarget = GitHubItemTarget & {
   href: string;
 };
 
-function decodePathSegment(value: string): string | null {
+export function decodeGitHubPathSegment(value: string): string | null {
   try {
     const decoded = decodeURIComponent(value).trim();
     return decoded && decoded !== "." && decoded !== ".." ? decoded : null;
@@ -22,10 +22,10 @@ function decodePathSegment(value: string): string | null {
   }
 }
 
-function parseGitHubItemPath(url: URL): GitHubItemTarget | null {
+export function parseGitHubItemPath(url: URL): GitHubItemTarget | null {
   const segments = url.pathname.split("/").filter(Boolean);
-  const owner = decodePathSegment(segments[0] ?? "");
-  const repo = decodePathSegment(segments[1] ?? "");
+  const owner = decodeGitHubPathSegment(segments[0] ?? "");
+  const repo = decodeGitHubPathSegment(segments[1] ?? "");
   const surface = segments[2];
   const numberText = segments[3] ?? "";
   if (!owner || !repo || !/^[1-9]\d{0,9}$/.test(numberText)) {
@@ -52,35 +52,12 @@ export function parseGitHubLinkTarget(href: string): GitHubLinkTarget | null {
   return target ? { ...target, href: url.href } : null;
 }
 
-export function formatGitHubLinkLabel(url: URL): string {
-  const segments = url.pathname.split("/").filter(Boolean);
-  const item = parseGitHubItemPath(url);
-  if (item && segments.length === 4 && !url.search && !url.hash) {
-    return `#${item.number}`;
-  }
-  if (item) {
-    return url.href;
-  }
-  if (segments.length === 2) {
-    return segments.map((segment) => decodePathSegment(segment) ?? segment).join("/");
-  }
-  if (segments[2] === "blob" && segments.length > 4) {
-    const filename = decodePathSegment(segments.at(-1) ?? "");
-    if (filename) {
-      return filename;
-    }
-  }
-  const fallbackSegments = segments.length > 2 ? segments.slice(2) : segments;
-  const path = fallbackSegments.map((segment) => decodePathSegment(segment) ?? segment);
-  return ["github.com", ...path].join("/");
-}
-
 export function gitHubProfileUrl(login: string): string {
   return `https://${GITHUB_HOST}/${encodeURIComponent(login)}`;
 }
 
-// Built from the parsed parts rather than the source href, which isGitHubItemRootPath
-// shows may already carry its own sub-path, query, or comment fragment.
+// Build from parsed parts because the source href may already carry its own
+// sub-path, query, or comment fragment.
 export function gitHubFilesChangedUrl(target: GitHubItemTarget): string {
   const repoPath = `${encodeURIComponent(target.owner)}/${encodeURIComponent(target.repo)}`;
   return `https://${GITHUB_HOST}/${repoPath}/pull/${target.number}/files`;
